proxy/nss/confparse: fix test heap corruption
Background:
The confparse test suite was passing its assertions but then crashing in
GTest XML teardown. The real issue was a use-after-free in the test
fixture, plus a latent uint8 adapter bug in the parser itself.

Specifically:
- Reset result.match to NULL after freeing it in confparse_test so the
  next parse_buffer() call does not feed a dangling pointer back into
  reallocarray().
- Fix adapter_uint8() to store through a uint8_t pointer instead of a
  uint16_t pointer.

+++ proxy/nss/confparse/confparse.c
@@ -190,7 +190,7 @@ error_code_e adapter_uint8(parse_context_t *ctx, const char *start,
 {
 	uint64_t value = 0;
 	int status = parse_uint64(ctx, UINT8_MAX, &value);
-	*(uint16_t *)((char *)(dest) + data->offset) = value;
+	*(uint8_t *)((char *)(dest) + data->offset) = value;
 	return status;
 }
